# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legal-AI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "parameters": [
          {
            "name": "x-anthropic-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anthropic-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/summary": {
      "post": {
        "summary": "Summary",
        "operationId": "summary_api_summary_post",
        "parameters": [
          {
            "name": "x-anthropic-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Anthropic-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/checklists/{crime_type}": {
      "get": {
        "summary": "Checklist",
        "operationId": "checklist_api_checklists__crime_type__get",
        "parameters": [
          {
            "name": "crime_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Crime Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checklist"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{name}": {
      "get": {
        "summary": "Template",
        "operationId": "template_api_templates__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/support/match": {
      "post": {
        "summary": "Support Match",
        "description": "Match answers to support programs. A program matches when every declared\ncriterion is satisfied; programs with no criteria always apply.",
        "operationId": "support_match_api_support_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EligibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/feedback": {
      "post": {
        "summary": "Feedback",
        "description": "Negative feedback becomes an eval candidate. Only the question text is\nkept — never the answer, and never anything the user typed about themselves\nbeyond the question they chose to send.",
        "operationId": "feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/client-error": {
      "post": {
        "summary": "Client Error",
        "description": "Frontend crash reports (no user content). Grep logs for client_error to\ncompute the crash-free session metric.",
        "operationId": "client_error_api_client_error_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientError"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "lang": {
            "type": "string",
            "enum": [
              "ko",
              "en"
            ],
            "title": "Lang",
            "default": "ko"
          },
          "simple": {
            "type": "boolean",
            "title": "Simple",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest"
      },
      "Checklist": {
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "urgent": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "type": "array",
            "title": "Urgent",
            "default": []
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "minItems": 1,
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "label",
          "items",
          "sources"
        ],
        "title": "Checklist"
      },
...
      "ClientError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "stack": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ClientError"
      },
...
      "EligibilityRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "EligibilityRequest"
      },
      "Feedback": {
        "properties": {
          "helpful": {
            "type": "boolean",
            "title": "Helpful"
          },
          "question": {
            "type": "string",
            "maxLength": 2000,
            "title": "Question",
            "default": ""
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "helpful"
        ],
        "title": "Feedback"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceLink": {
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "label",
          "url"
        ],
        "title": "SourceLink"
      },
...
      "SummaryRequest": {
        "properties": {
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "SummaryRequest"
      },
      "SupportMatchResponse": {
        "properties": {
          "matched": {
            "items": {
              "$ref": "#/components/schemas/SupportProgram"
            },
            "type": "array",
            "title": "Matched"
          },
          "others": {
            "items": {
              "$ref": "#/components/schemas/SupportProgram"
            },
            "type": "array",
            "title": "Others"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "title": "Sources"
          },
          "verified_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified On"
          }
        },
        "type": "object",
        "required": [
          "matched",
          "others",
          "notes",
          "disclaimer",
          "sources"
        ],
        "title": "SupportMatchResponse"
      },
...
    }
  }
}